change id type to bigint

+++ sql/create-tables.sql
@@ -1,5 +1,5 @@
 create table artist (
-    id int primary key auto_increment,
+    id bigint primary key auto_increment,
     email nvarchar(125) unique not null,
     password varchar(128) not null,
     nickname nvarchar(100) unique not null,
@@ -7,8 +7,8 @@ create table artist (
 );
 
 create table illustration(
-    id int primary key auto_increment,
-    artistId int,
+    id bigint primary key auto_increment,
+    artistId bigint,
     name nvarchar(200) not null,
     imageUrl nvarchar(2048) not null,
     createdAt datetime not null,
